task: Add Mac OS X backend (#3414787)
As of October 2011, Mac OS X does not support named semaphores.
So instead of trying to rework the existing Linux code to work
with semaphores in both OS X and Linux, it's easier to reimplement
with pthread mutexes for just OS X.

From rogerman.

xrmx: fixed gcc warnings for storing NULL in a pthread_t.
Also while at it moved the includes on top.


git-svn-id: svn://svn.code.sf.net/p/desmume/code@4124 5577840e-afb3-4e59-9ef3-2b7f898b1952